+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## Enhancements
 * Adds support for `NotificationConfigurationSubscribableChoice` type `Project` to support Project Level Notifications by @jillirami [#1350](https://github.com/hashicorp/go-tfe/pull/1350)
+* Adds `Explorer` to support querying organization data across workspaces and exporting it as CSV by @a9logic [#1356](https://github.com/hashicorp/go-tfe/pull/1356)
 
 ## Bug Fixes
 * Adds support for s390x tool versions, which were previously blocked by @JenGoldstrich [#1357](https://github.com/hashicorp/go-tfe/pull/1357)